many updates...

+++ src/printf.h
@@ -113,7 +113,7 @@ int fctprintf(void (*out)(char character, void* arg), void* arg, const char* for
 }
 #endif
 
-void bprintn(uint64_t n);
+void bprintn(uint8_t n);
 
 
 #endif  // _PRINTF_H_
